Add customizable OverviewHeaderAction
